fix(design-system): point the chart tests at the moved util
The move to utils/chart.ts missed three files under packages/design-system/
tests/, which the app's tsconfigs do not cover -- so ui/ type-checked clean
while the design-system CI job failed to resolve the old path.
